driver/requirements.txt
Flask==0.10.1
Flask-HTTPAuth==2.3.0
flask-marshmallow==0.6.2
Flask-Script==2.0.5
marshmallow==2.6.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
itsdangerous==0.24
six==1.9.0
pyusb
